# swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: TelPro - FileMaker App (Hive)
# during dev, should point to your local machine
#host: 52.34.207.5:5058
host: localhost:5000
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /api/addCompanyEmailTemplate:
    # binds a127 app logic to a route
    x-swagger-router-controller: emailTemplate_ctrl
    post:
      tags: ['Email Template Module']
      description: Add Company Email Template 
      summary: Add Company Email Template
      # used as the method name of the controller
      operationId: addCompanyEmailTemplate
      parameters:
        - name: body
          in: body
          description: Add Company Email Template
          schema:
           $ref: '#/definitions/addCompanyEmailTemplate'
          required: true        
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse" 
  /api/getCompanyEmailTemplateDetails:
    # binds a127 app logic to a route
    x-swagger-router-controller: emailTemplate_ctrl
    post:
      tags: ['Email Template Module']
      description: Get Company Email Template Details 
      summary: Get Company Email Template Details
      # used as the method name of the controller
      operationId: getCompanyEmailTemplateDetails
      parameters:
        - name: body
          in: body
          description: Get Company Email Template Details
          schema:
           $ref: '#/definitions/getCompanyEmailTemplateDetails'
          required: true        
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /api/updateCompanyEmailTemplate:
    # binds a127 app logic to a route
    x-swagger-router-controller: emailTemplate_ctrl
    post:
      tags: ['Email Template Module']
      description: Update Company Email Template 
      summary: Update Company Email Template 
      # used as the method name of the controller
      operationId: updateCompanyEmailTemplate
      parameters:
        - name: body
          in: body
          description: Update Company Email Template
          schema:
           $ref: '#/definitions/updateCompanyEmailTemplate'
          required: true        
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /api/getCompanyEmailTemplates:
    # binds a127 app logic to a route
    x-swagger-router-controller: emailTemplate_ctrl
    get:
      tags: ['Email Template Module']
      description: get edit template list
      summary: get edit template list
      # used as the method name of the controller
      operationId: getCompanyEmailTemplates
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /api/getItemConfiglist:
    # binds a127 app logic to a route
    x-swagger-router-controller: estimateConfiguration_ctrl
    post:
      tags: ['Estimate Configuration Module']
      description: get Item Config list
      summary: get Item Config list
      # used as the method name of the controller
      operationId: getItemConfiglist
      parameters:
        - name: body
          in: body
          description: get get Item Configlist
          schema:
           $ref: '#/definitions/getItemConfiglist'
          required: true        
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /api/generateEstimatePdf:
    # binds a127 app logic to a route
    x-swagger-router-controller: estimateConfiguration_ctrl
    post:
      tags: ['Estimate Configuration Module']
      description: genrate estimate item  pdf
      summary: genrate estimate item  pdf
      # used as the method name of the controller
      operationId: generateEstimatePdf
      parameters:
        - name: body
          in: body
          description: genrate estimate item  pdf
          schema:
           $ref: '#/definitions/generateEstimatePdf'
          required: true        
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /api/addEstimateFormat:
    # binds a127 app logic to a route
    x-swagger-router-controller: estimateConfiguration_ctrl
    post:
      tags: ['Estimate Configuration Module']
      description:  Add format for estimate
      summary:  Add format for estimate
      # used as the method name of the controller
      operationId: addEstimateFormat
      parameters:
        - name: body
          in: body
          description:  Add format for estimate
          schema:
           $ref: '#/definitions/addEstimateFormat'
          required: true        
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /api/updateEstimateFormat:
    # binds a127 app logic to a route
    x-swagger-router-controller: estimateConfiguration_ctrl
    post:
      tags: ['Estimate Configuration Module']
      description:  use to update format for estimate
      summary:  use to update format for estimate
      # used as the method name of the controller
      operationId: updateEstimateFormat
      parameters:
        - name: body
          in: body
          description:  use to update format for estimate
          schema:
           $ref: '#/definitions/updateEstimateFormat'
          required: true        
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
...
  appResponse:
    required:
      - code
    properties:
      code:
        type: number
      message:
        type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
...
  addCompanyEmailTemplate:
    type: object
    properties: 
      subject:
        type: string
      title:
        type: string
      description:
        type: string
      code:
        type: string 
  getCompanyEmailTemplateDetails:
    type: object
    properties: 
      emailtemplateId:
        type: string              
  updateCompanyEmailTemplate:
    type: object
    properties: 
      emailtemplateId:
        type: string
      subject:
        type: string
      title:
        type: string
      description:
        type: string
      code:
        type: string
...
  getItemConfiglist:
    type: object
    properties:
      estimateId:
        type: string
  generateEstimatePdf:
    type: object
    properties:
      estimateId:
        type: string
      htmlContent:
        type: string
  addEstimateFormat:
    type: object
    properties:
      estimateId:
        type: string
  updateEstimateFormat:
    type: object
    properties:
      configId:
        type: string
...